requirements.txt
graphene==3.1.1
graphql-core==3.2.3
wheel~=0.36.2
setuptools~=57.1.0
numpy~=1.23.3
jsonschema~=3.2.0
pandas==1.5.0
openpyxl>=2.5.0
networkx~=2.5.1
lxml~=4.9.1
requests~=2.25.1
iso8601~=0.1.14
chardet~=4.0.0
jinja2~=3.0.1
beautifulsoup4~=4.9.3
mzml2isa==1.0.3
biopython~=1.79
progressbar2~=3.53.1
deepdiff~=5.5.0
PyYAML~=5.4.1
bokeh~=2.3.2
certifi==2021.5.30
flake8==3.9.2
ddt==1.4.2
behave==1.2.6
httpretty==1.1.3
sure==2.0.0
coveralls~=3.1.0
rdflib~=6.0.2
SQLAlchemy~=1.4.42
python-dateutil~=2.8.1